Change QUEUE_TIMEOUT_GET to 1 second
The 0.1 is good for responsiveness, but this comes at the cost
of busy waiting.  1 second is the value used in boto's glacier
thread pools, and that gives us a good tradeoff between responsiveness
and cpu usage.

Before this change, just starting the executor cause the python process
to sit at around 30% cpu (waiting for tasks to be added to the work
queue).  With this change the cpu sits around 3% after starting the
executor.