Add support for per sub-command aliases
The [CLI alias feature](https://docs.aws.amazon.com/cli/latest/userguide/cli-usage-alias.html)
currently supports creating aliases as top level commands, at the
same level that service commands currently reside.  For example,
given this alias file:

```
[toplevel]
iam-roles = iam list-roles
```

you can run `aws iam-roles`, which will be equivalent to
`aws iam list-roles`.  This helps reduce the amount of typing for
commonly used commands.

However, this requires all aliases be in the to level namespace,
which has hundreds of services.  This makes it hard to remember
and look up what aliases you have available.

This PR adds support for creating aliases within namespaces of
existing commands/subcommands.

There is some implementation specific details with respect to
event names that we needed to rely onin order to make aliases of nested
commands work (i.e. `aws foo bar <myalias>`).

The implementation is different from the approach taken by
the `[toplevel]` aliases for several reasons:

* The toplevel aliases make the assumption that a single parser can be
  used for any of the alias values.  This is essentially the "global
  args" parser, so this holds true for top level aliase but isn't
  guaranteed for subcommands, custom commands, etc.
* We need to ensure we scope the alias expansion to the current
  namespace we're traversing so we need an approach that enforces this.
* There may be additional logic that happens when invoking a command,
  even within a given namespace.  The safer option is to re-execute
  all this logic after expanding the alias.

So the high level idea is to take the reference to the parent command
object, and reinvoke it with the expanded alias value, after
picking off, and updating, and global parameters provided in the
alias value.

This approach also has the benefit of not modifying the existing
code (except for extract pieces out into a shared base class) so
it's conservative in terms of the likelihood of
breaking existing aliases.

Closes #7386